+++ fixplz-web/src/admin/assets/scss/main-page.scss
