+++ fixplz-web/src/admin/assets/scss/stat-page.scss
